refactor(team-builder): extract useIdentityState hook from IdentityLaneReal
Move all derived data (types, gender, level, mega flags, error
partitions) plus handlers (nick blur, gender toggle, shiny toggle,
species pick) and nickname draft state from IdentityLaneReal into a
dedicated useIdentityState hook in lanes/identity/use-identity-state.ts.

Pure code-move with no behavior change. IdentityLaneReal now calls the
hook and destructures its return instead of local variables. Picker-open
useState calls stay in IdentityLaneReal — those move into individual
cells in a later step.

Also removed the now-unused nextGender helper from identity-lane.tsx
and moved errorsForField/errorsForFields into the hook file where they
are used.

Setting up the file structure for the upcoming three-layout split:
SingleRow, MidStack, Vertical will all consume this same hook.
